Moved education to its own folder
