+++ client/src/components/Education/educationCreate.js
@@ -1,10 +1,10 @@
 import React, { Component } from "react";
-import Sidebar from "./subComponents/sidebar";
+import Sidebar from "../subComponents/sidebar";
 import axios from "axios";
-import Navbar from "./subComponents/navbar";
+import Navbar from "../subComponents/navbar";
 import { Redirect } from "react-router-dom";
 
-const urls = require("../config/config.json");
+const urls = require("../../config/config.json");
 
 class EducationCreate extends Component {
   constructor(props) {
